IWYU support (part 6): more misc changes (#29691)
* OSX-specific include in translations.cpp

* Move M_PI workaround to math_defines.h

Turns out M_PI isn't standard C, it's POSIX.  In particular, it's not
present on Mingw.  So we need to provide our own fallback definition.

* Include platform_win.h in main.cpp

Needed for GetUserDefaultLCID.

* Update list of pending IWYU PRs

* Stop cyclic include between player.h and stomach.h

* Prevent player.h including game.h

IWYU would force player.h to include game.h to get some type ids, which
causes big cyclic include issues.  Add those type declarations to
player.h to avoid that.

* Move game_message_type to enums.h

This reduces the extent to which output.h need be included (which is a
fairly expensive header).
