+++ doc/DEVELOPER_TOOLING.md
@@ -16,11 +16,10 @@ See the [JSON style guide](JSON_STYLE.md).
 required headers and add and remove includes as appropriate.
 
 Running on this codebase revealed some issues.  You will need a version of IWYU
-where the following PRs have been merged (which has not yet happened at time of
+where the following PR has been merged (which has not yet happened at time of
 writing):
 
-* https://github.com/include-what-you-use/include-what-you-use/pull/637
-* https://github.com/include-what-you-use/include-what-you-use/pull/641
+* https://github.com/include-what-you-use/include-what-you-use/pull/681
 
 Once you have IWYU built, build the codebase using cmake, with
 `CMAKE_EXPORT_COMPILE_COMMANDS=ON` on to create a compilation database
