+++ src/enums.h
@@ -383,4 +383,23 @@ enum class distraction_type {
     weather_change,
 };
 
+enum game_message_type : int {
+    m_good,    /* something good happened to the player character, e.g. health boost, increasing in skill */
+    m_bad,      /* something bad happened to the player character, e.g. damage, decreasing in skill */
+    m_mixed,   /* something happened to the player character which is mixed (has good and bad parts),
+                  e.g. gaining a mutation with mixed effect*/
+    m_warning, /* warns the player about a danger. e.g. enemy appeared, an alarm sounds, noise heard. */
+    m_info,    /* informs the player about something, e.g. on examination, seeing an item,
+                  about how to use a certain function, etc. */
+    m_neutral,  /* neutral or indifferent events which aren’t informational or nothing really happened e.g.
+                  a miss, a non-critical failure. May also effect for good or bad effects which are
+                  just very slight to be notable. This is the default message type. */
+
+    m_debug, /* only shown when debug_mode is true */
+    /* custom SCT colors */
+    m_headshot,
+    m_critical,
+    m_grazing
+};
+
 #endif
