+++ src/math_defines.h
@@ -15,4 +15,10 @@
 #define _USE_MATH_DEFINES
 #include <cmath>
 
+// And on mingw even that doesn't work, so we are forced to have our own
+// fallback definition.
+#ifndef M_PI
+#define M_PI 3.14159265358979323846
+#endif
+
 #endif // CATA_MATH_DEFINES_H
